requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-core
faiss-cpu
openai
tiktoken
python-dotenv
sentence-transformers
chromadb
langchain-chroma
requests
biopython
torch
scikit-learn
fuzzywuzzy
python-Levenshtein
